address greptile review: per-provider deficit notch, startedAt filter, weekRange refresh, deduplicate providerDisplayName
